requirements.txt
torch==2.1.2
numpy==1.26.4
Pillow==9.5.0
transformers== 4.42.3
bitsandbytes==0.43.3
accelerate==0.32.1
ultralytics==8.2.79
